Employing permissions and roles  in the project

+++ app/Http/Controllers/CategoryController.php
@@ -6,12 +6,21 @@ use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use App\Models\Course;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Spatie\Permission\Traits\HasRoles;
 
 class CategoryController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
+    use HasRoles;
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware('check_user_role');
+    }
+
     public function index()
     {
         $categories=Category::all();
@@ -32,14 +41,19 @@ class CategoryController extends Controller
      */
     public function store(CategoryRequest $request)
     {
-        $validatedData=$request->validated();
-        if ($request->hasFile('image')) {
-            $file=$request->file('image');
-            $path = UploadImage($file, "categories",'category');
-            $validatedData['image'] = $path;
+        if($request->user()->hasRole('admin')){
+            $validatedData=$request->validated();
+            if ($request->hasFile('image')) {
+                $file=$request->file('image');
+                $path = UploadImage($file, "categories",'category');
+                $validatedData['image'] = $path;
+            }
+            Category::create($validatedData);
+            return redirect()->route('categories.index')->with('success', 'Category created successfully!');
         }
-        Category::create($validatedData);
-        return redirect()->route('categories.index')->with('success', 'Category created successfully!');
+        else{
+            abort(403,'you do not have permissions to add a Category');
+        }
     }
 
     /**
@@ -63,18 +77,22 @@ class CategoryController extends Controller
      */
     public function update(CategoryRequest $request, Category $category)
     {
-        $validatedData=$request->validated();
-        if($request->hasFile('image')){
-            $file = $request->file('image');
-            $path = UploadImage($file,'categories','category');    
-            $validatedData['image'] = $path;
-        }
-        $category->update([
-            'name' => $validatedData['name'],
-            'image' => $validatedData['image'],
+        if($request->user()->hasRole('admin')){
+            $validatedData=$request->validated();
+            if($request->hasFile('image')){
+                $file = $request->file('image');
+                $path = UploadImage($file,'categories','category');    
+                $validatedData['image'] = $path;
+            }
+            $category->update([
+                'name' => $validatedData['name'],
+                'image' => $validatedData['image'],
             ]);
-        return redirect()->route('categories.index')->with('success', 'Category updated successfully!');
-
+            return redirect()->route('categories.index')->with('success', 'Category updated successfully!');
+        }
+        else{
+            abort(403,'you do not have permissions to update a Category');
+        }
     }
 
     /**
@@ -82,7 +100,13 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
-        return redirect()->route('categories.index')->with('success', 'Category deleted successfully!');
+        $user=Auth::user();
+        if($user->hasRole('admin')){
+            $category->delete();
+            return redirect()->route('categories.index')->with('success', 'Category deleted successfully!');
+        }
+        else{
+            abort(403,'you do not have permissions to delete a Category');
+        }
     }
 }
